Backed out changeset ee2d0bbca72c (bug 1159632) for windows xpcshell permafail
UltraBlame original commit: 5a28a5474ba69778532f4f715b5b1213a5d4e5e4

+++ toolkit/components/jsdownloads/test/unit/head.js
@@ -137,15 +137,8 @@ function getTempFile(aLeafName)
   do_check_false(file.exists());
 
   do_register_cleanup(function () {
-    try {
+    if (file.exists()) {
       file.remove(false);
-    } catch (ex if (ex instanceof Components.Exception &&
-                    (ex.result == Cr.NS_ERROR_FILE_ACCESS_DENIED ||
-                     ex.result == Cr.NS_ERROR_FILE_TARGET_DOES_NOT_EXIST))) {
-        
-        
-        
-        
     }
   });
